Fix support for renamed files, aka git mv in upload.py.
Prior this change, a renamed file would only have the new file listed on
rietveld and not the previous one. This is caused by the default behavior of
"git diff".

This change calls "git diff" two times to extract all the necessary data.

---

The review was at http://codereview.appspot.com/4333051. This code is
incomplete; GitVCS.GetBaseFile() doesn't use --rev while GitVCS.GenerateDiff()
does, so while this patch works for chromium, it's unsuitable for general use.